isDiscounted compares 2 strings instead of 2 numbers
While working with the demo-store template, I was having issues getting the discounts to show up.

My product:
- price: 50.0
- compareAtPrice: 100.0

This is an item that I was expecting to show a sale price. I loaded the demo-store and was unable to see the sale. I debugged the issue to the utils isDiscounted() function.

```
export function isDiscounted(price: MoneyV2, compareAtPrice: MoneyV2) {
  if (compareAtPrice?.amount > price?.amount) {
    return true;
  }
  return false;
}
```

We see that we are comparing a MoneyV2.amount with another MoneyV2.amount. What's wrong with this?

```
export declare type MoneyV2 = {
    __typename?: 'MoneyV2';
    /** Decimal money amount. */
    amount: Scalars['Decimal'];
    /** Currency of the money. */
    currencyCode: CurrencyCode;
};
```

In the type definition, we see that amount: Scalars['Decimal']. However, when I access the attribute, I actually am getting back a string representation of a decimal.

So for my example
```
compareAtPrice?.amount > price?.amount
// "100.0" > "50.0"
// => false
```

Instead, let's cast the strings to their decimals and compare.